+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + Inline metas more eagerly (#178)
 + Improve pest utility
 + Unification now unifies extension
++ Fix expand-global on neutral terms
 
 # 0.2.2